requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
APScheduler==3.10.4
openai==1.14.2
anthropic==0.21.3
httpx==0.27.0
Pillow==10.1.0
requests==2.31.0
